% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliable_grayscale.R
\name{ensure_grayscale}
\alias{ensure_grayscale}
\title{Convert cimg object to grayscale via XYZ}
\usage{
ensure_grayscale(img, bg = 1)
}
\arguments{
\item{img}{An \code{\link[imager]{cimg}} object containing an image (from imager)}

\item{bg}{Background color (default is white)}
}
\description{
Improves on the XYZ method of \code{\link[imager]{grayscale}}
by handling alpha channel and retaining relative luminance of
the image
}
