% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ort-package.R
\docType{package}
\name{ort-package}
\alias{ort-package}
\title{ort: Create a Data Frame Representation of an Image}
\description{
The package takes images imported via imager and outputs a data frame that, when plotted in base R, looks like the original image.
}
\author{
\strong{Maintainer}: Ethan C. Brown \email{ethancbrown@gmail.com} (Maintainer)

Authors:
\itemize{
  \item Falling Pineapples (Original author)
}

}
\keyword{internal}
