% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o.R
\name{ort}
\alias{ort}
\title{Create a data frame representaiton of an image}
\usage{
ort(
  img,
  bg = 1,
  grid_size = function(x) {
     1/(250 * (1.0001 - x))
 },
  grid_offset = function(x) {
     sin(x * 100) * 10
 }
)
}
\arguments{
\item{img}{An \code{\link[imager]{cimg}} object containing an image (from imager)}

\item{bg}{Background color (default is white)}

\item{grid_size}{Size of grid used for conversion}

\item{grid_offset}{Offset of grid used for conversion}
}
\value{
A data frame with two columns, `x` and `y` that
hopefully resembles the image when plotted
}
\description{
Takes a \code{\link[imager]{cimg}} object and
attempts to create a data frame that, when plotted using the default
\code{\link[graphics]{plot}} in base R, looks like the original image.
}
