% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ort-package.R
\docType{package}
\name{ort-package}
\alias{ort-package}
\title{ort: Create a Data Frame Representation of an Image}
\description{
Takes images, imported via 'imager', and converts them into a data frame that can be plotted to look like the imported image. This can be used for creating data that looks like a specific image. Additionally, images with color and alpha channels can be converted to grayscale in preparation for converting to the data frame format.
}
\author{
\strong{Maintainer}: Ethan C. Brown \email{ethancbrown@gmail.com} (Maintainer)

Authors:
\itemize{
  \item Falling Pineapples (Original author)
}

}
\keyword{internal}
